Fix zoom for track pad
