+++ app/src/routes/project/Canvas/PanOverlay.tsx
@@ -1,7 +1,6 @@
 import { EditorMode } from '@/lib/models';
 import clsx from 'clsx';
 import { observer } from 'mobx-react-lite';
-import { useCallback, useEffect, useState } from 'react';
 import { useEditorEngine } from '..';
 
 interface PanOverlayProps {
@@ -11,50 +10,11 @@ interface PanOverlayProps {
             y: number;
         }>
     >;
+    isPanning: boolean;
+    setIsPanning: React.Dispatch<React.SetStateAction<boolean>>;
 }
-const PanOverlay = observer(({ setPosition }: PanOverlayProps) => {
+const PanOverlay = observer(({ setPosition, isPanning, setIsPanning }: PanOverlayProps) => {
     const editorEngine = useEditorEngine();
-    const [isPanning, setIsPanning] = useState(false);
-
-    const spaceBarDown = (e: KeyboardEvent) => {
-        if (e.key === ' ') {
-            editorEngine.mode = EditorMode.PAN;
-        }
-    };
-
-    const spaceBarUp = useCallback((e: KeyboardEvent) => {
-        if (e.key === ' ') {
-            editorEngine.mode = EditorMode.DESIGN;
-        }
-    }, []);
-
-    const middleMouseButtonDown = (e: MouseEvent) => {
-        if (e.button === 1) {
-            editorEngine.mode = EditorMode.PAN;
-            setIsPanning(true);
-        }
-    };
-
-    const middleMouseButtonUp = (e: MouseEvent) => {
-        if (e.button === 1) {
-            editorEngine.mode = EditorMode.DESIGN;
-            setIsPanning(false);
-        }
-    };
-
-    useEffect(() => {
-        window.addEventListener('keydown', spaceBarDown);
-        window.addEventListener('keyup', spaceBarUp);
-        window.addEventListener('mousedown', middleMouseButtonDown);
-        window.addEventListener('mouseup', middleMouseButtonUp);
-
-        return () => {
-            window.removeEventListener('keydown', spaceBarDown);
-            window.removeEventListener('keyup', spaceBarUp);
-            window.removeEventListener('mousedown', middleMouseButtonDown);
-            window.removeEventListener('mouseup', middleMouseButtonUp);
-        };
-    }, []);
 
     const startPan = (event: React.MouseEvent<HTMLDivElement>) => {
         event.preventDefault();
